feat: add settings panel component
